requirements.txt
action-tutorials-py==0.9.4
ament-copyright==0.9.8
ament-cppcheck==0.9.8
ament-cpplint==0.9.8
ament-flake8==0.9.8
ament-index-python==1.1.0
ament-lint==0.9.8
ament-lint-cmake==0.9.8
ament-package==0.9.5
ament-pep257==0.9.8
ament-uncrustify==0.9.8
ament-xmllint==0.9.8
anfis==0.3.1
argcomplete==1.8.1
asttokens==2.2.1
async-generator==1.10
atomicwrites==1.1.5
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.2.0
beautifulsoup4==4.8.2
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.2.0
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
cbor==1.0.0
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
click==8.1.3
cloud-init==23.1.1
colcon-argcomplete==0.3.3
colcon-bash==0.4.2
colcon-cd==0.1.1
colcon-cmake==0.2.27
colcon-common-extensions==0.3.0
colcon-core==0.12.1
colcon-defaults==0.2.8
colcon-devtools==0.2.3
colcon-installed-package-information==0.1.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.2.15
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.3.3
colcon-package-selection==0.2.10
colcon-parallel-executor==0.2.4
colcon-pkg-config==0.1.0
colcon-powershell==0.3.7
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.1
colcon-ros==0.3.23
colcon-test-result==0.3.8
colcon-zsh==0.4.0
colorama==0.4.3
comm==0.1.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cov-core==1.15.0
coverage==4.5.2
cryptography==38.0.3
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
demo-nodes-py==0.9.4
distlib==0.3.6
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
domain-coordinator==0.9.2
empy==3.3.2
entrypoints==0.3
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.0
firebase-admin==6.0.1
flake8==3.7.9
Flask==2.3.2
Flask-Cors==3.0.10
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
h11==0.14.0
html5lib==1.0.1
httplib2==0.21.0
hyperlink==19.0.0
idna==2.8
ifcfg==0.18
importlib-metadata==6.3.0
incremental==16.10.1
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
lark-parser==0.8.1
launch==0.10.10
launch-ros==0.11.7
launch-testing==0.10.10
launch-testing-ros==0.11.7
launch-xml==0.10.10
launch-yaml==0.10.10
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linecache2==1.0.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.1.3
matplotlib==3.1.2
matplotlib-inline==0.1.6
mccabe==0.6.1
mock==3.0.5
more-itertools==4.2.0
mpi4py==3.0.3
msgpack==1.0.4
mypy==0.761
mypy-extensions==0.4.3
mysql-connector-python==8.0.33
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==3.0
nose2==0.9.1
notify2==0.3
numpy==1.24.2
oauthlib==3.1.0
olefile==0.46
osrf-pycommon==0.1.11
outcome==1.2.0
packaging==20.3
pandas==2.0.0
parso==0.8.3
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==3.2.0
plotly==5.15.0
pluggy==0.13.0
# Editable install with no version control (pohm-robotics-arm==0.0.0)
-e /home/alexpc/dev_ws/build/pohm_robotics_arm
prompt-toolkit==3.0.38
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.5.1
pure-eval==0.2.2
py==1.8.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.21
pycups==1.9.73
pydocstyle==2.1.1
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.15.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.6.0
pymacaroons==0.13.0
pymodbus==3.3.0
PyNaCl==1.3.0
pyOpenSSL==22.1.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyqtgraph==0.13.3
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pytest==4.6.9
pytest-cov==2.8.1
pytest-mock==1.10.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-snappy==0.5.3
PyTrie==0.2
pytz==2023.3
PyYAML==5.3.1
pyzmq==25.0.2
quality-of-service-demo-py==0.9.4
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
ros2action==0.9.13
ros2bag==0.3.10
ros2cli==0.9.13
ros2component==0.9.13
ros2doctor==0.9.13
ros2interface==0.9.13
ros2launch==0.11.7
ros2lifecycle==0.9.13
ros2multicast==0.9.13
ros2node==0.9.13
ros2param==0.9.13
ros2pkg==0.9.13
ros2run==0.9.13
ros2service==0.9.13
ros2topic==0.9.13
rosdep==0.18.0
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosidl-runtime-py==0.9.1
rospkg==1.5.0
rospkg-modules==1.5.0
rpyutils==0.2.0
rqt-action==1.0.2
rqt-console==1.1.2
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-msg==1.1.1
rqt-plot==1.1.1
rqt-publisher==1.3.0
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-top==1.0.2
rqt-topic==1.3.0
rsa==4.9
scikit-fuzzy==0.4.2
scipy==1.10.1
SecretStorage==2.3.1
selenium==4.9.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.0
snowballstemmer==2.0.0
sortedcontainers==2.4.0
sos==4.4
soupsieve==1.9.5
sros2==0.9.5
ssh-import-id==5.10
stack-data==0.6.2
systemd-python==234
teleop-twist-keyboard==2.3.2
tenacity==8.2.2
testresources==2.0.0
topic-monitor==0.9.4
tornado==6.2
traceback2==1.4.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==18.9.0
txaio==2.10.0
typed-ast==1.4.1
typing_extensions==4.5.0
tzdata==2023.3
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
uritemplate==4.1.1
urllib3==2.0.2
vcstool==0.3.0
virtualenv==20.23.0
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.3.6
wsaccel==0.6.2
wsproto==1.2.0
zipp==1.0.0
zope.interface==4.7.1
